move Split::IntoMultipleColumns transform class to separate file
